feat: 架构升级  JWT认证 + 共享订阅 + Redis Streams + TimescaleDB + 运维配套
Phase 1: EMQX内置JWT认证 + Flutter端密码切换为JWT + 证书校验加固
Phase 2: \/共享订阅 + HubRedis共享状态 + Session清理 + 密码脱敏
Phase 3: Redis Streams消费者(消费组/ACK/死信) + EMQX Rule Engine SQL参考
Phase 4: TimescaleDB超表迁移脚本(自动压缩+连续聚合)
Phase 5: Docker Compose + K8s HPA + Metrics端点 + MQTT压测脚本 + Prometheus告警

+++ inv_device_server/internal/model/device.go
@@ -117,7 +117,7 @@ type EnergyData struct {
 	TotalDischarge  float64 `json:"total_discharge"`
 	DailyLoad       float64 `json:"daily_load"`
 	TotalLoad       float64 `json:"total_load"`
-	RuntimeHours    int     `json:"runtime_hours"`
+	RuntimeHours    float64 `json:"runtime_hours"`
 
 	SN         string    `json:"-"`
 	ReceivedAt time.Time `json:"-"`
